Move template hierarchy utilities to a shared file
